fastapi-background-jobs/pyproject.toml
[project]
name = "fastapi-background-jobs"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
  "fastapi",
  "uvicorn[standard]",
  "sqlalchemy",
  "psycopg2-binary",
  "pydantic-settings",
  "python-dotenv",
  "alembic",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
